Refine locale preference normalization flow
