+++ app/src/main/java/com/example/alias/AliasApp.kt
@@ -1,8 +1,6 @@
 package com.example.alias
 
 import android.app.Application
-import androidx.appcompat.app.AppCompatDelegate
-import androidx.core.os.LocaleListCompat
 import com.example.alias.data.settings.SettingsRepository
 import dagger.hilt.android.HiltAndroidApp
 import javax.inject.Inject
@@ -49,19 +47,6 @@ class AliasApp : Application() {
                     }
                 }
         }
-    }
-
-    private fun resolveInitialLocalePreference(stored: String): String {
-        val appLocales = AppCompatDelegate.getApplicationLocales()
-        val defaultLocales = LocaleListCompat.getAdjustedDefault()
-        val defaultTag = canonicalizeLocalePreference(defaultLocales.toLanguageTags())
-        val canonicalAppTag = canonicalizeLocalePreference(appLocales.toLanguageTags())
-        val appTag = when {
-            appLocales.isEmpty -> "system"
-            canonicalAppTag == defaultTag -> "system"
-            else -> canonicalAppTag
-        }
-        return if (appTag != "system") appTag else stored
     }
 }
